feat: chartImage view
